chore: refactor component lib
