+++ src/lib/filters/clear-filter-button.svelte
