+++ src/lib/filters/paginated-view.svelte
