+++ src/lib/filters/year-filter.svelte
@@ -1,5 +1,5 @@
 <script>
-    import OptionsList from "./options-list.svelte";
+    import OptionsList from "$lib/filters/options-list.svelte";
 
     let { startYear = $bindable(), endYear = $bindable(), preview } = $props();
