+++ src/lib/header.svelte
@@ -1,5 +1,5 @@
 <script>
-    import Nav from './nav.svelte';
+    import Nav from '$lib/nav.svelte';
 </script>
 
 <header>
