+++ src/routes/digital-catalog/+page.svelte
@@ -2,11 +2,11 @@
   	import { getBooks } from '$lib';
 	import Headercomponent from '$lib/header.svelte';
   	import Breadcrumbs from '$lib/breadcrumbs.svelte';
-  	import Search from '$lib/search.svelte';
-	import FiltersAside from '$lib/filters-aside.svelte';
-  	import FilterContainerSmall from '$lib/filter-container-small.svelte';
+  	import Search from '$lib/filters/search.svelte';
+	import FiltersAside from '$lib/filters/filters-aside.svelte';
+  	import FilterContainerSmall from '$lib/filters/filter-container-small.svelte';
   	import BookViewer from "$lib/book-viewer.svelte";
-  	import PaginatedView from '$lib/paginated-view.svelte';
+  	import PaginatedView from '$lib/filters/paginated-view.svelte';
   	import Footer from '$lib/footer.svelte';
   	import BookList from "$lib/book-list.svelte";
